refactor(chat): 컴포넌트를 기능별 폴더로 구조화
- ChatArea/, ChatInput/, ChatMessage/, Experience/, Layout/, Panel/ 폴더 생성
- 각 폴더에 index.ts 추가하여 export 관리
- 루트 index.ts를 폴더별 re-export로 변경
