+++ app/(main)/chat/[questionId]/_components/ChatArea/ChatAreaSkeleton.tsx
