+++ app/(main)/chat/[questionId]/_components/ChatArea/index.ts
@@ -1,0 +1,3 @@
+export { ChatAreaClient } from './ChatAreaClient';
+export { ChatAreaServer } from './ChatAreaServer';
+export { ChatAreaSkeleton } from './ChatAreaSkeleton';
