+++ app/(main)/chat/[questionId]/_components/ChatInput/index.ts
@@ -1,0 +1,2 @@
+export { ChatInput } from './ChatInput';
+export { InitialActionButton } from './InitialActionButton';
