+++ app/(main)/chat/[questionId]/_components/ChatMessage/AIAvatar.tsx
