+++ app/(main)/chat/[questionId]/_components/Experience/HashtagBadge.tsx
