+++ app/(main)/chat/[questionId]/_components/Layout/ChatProjectSummary.tsx
