+++ app/(main)/chat/[questionId]/_components/Layout/index.ts
@@ -1,0 +1,3 @@
+export { ChatPageShell } from './ChatPageShell';
+export { ChatProjectSummary } from './ChatProjectSummary';
+export { ChatQuestionTabs } from './ChatQuestionTabs';
