+++ app/(main)/chat/[questionId]/_components/Panel/index.ts
@@ -1,0 +1,4 @@
+export { SidePanelClient } from './SidePanelClient';
+export { ChatPanel } from './ChatPanel';
+export { ChatPanelTabs } from './ChatPanelTabs';
+export { DraftPanel } from './DraftPanel';
