+++ app/(main)/chat/[questionId]/_components/index.ts
@@ -1,15 +1,17 @@
-export { ChatPageShell } from './ChatPageShell';
-export { ChatAreaClient } from './ChatAreaClient';
-export { SidePanelClient } from './SidePanelClient';
-export { ChatAreaSkeleton } from './ChatAreaSkeleton';
-export { ExperienceCardsSkeleton } from './ExperienceCardsSkeleton';
-export { ChatProjectSummary } from './ChatProjectSummary';
-export { ChatQuestionTabs } from './ChatQuestionTabs';
-export { ChatPanel } from './ChatPanel';
-export { ChatPanelTabs } from './ChatPanelTabs';
-export { ExperienceListShell } from './ExperienceListShell';
-export { ExperienceCards } from './ExperienceCards';
-export { ExperienceCard } from './ExperienceCard';
-export { DraftPanel } from './DraftPanel';
-export { ChatMessageList } from './ChatMessageList';
-export { ChatInput } from './ChatInput';
+// ChatArea
+export * from './ChatArea';
+
+// ChatMessage
+export * from './ChatMessage';
+
+// ChatInput
+export * from './ChatInput';
+
+// Experience
+export * from './Experience';
+
+// Panel
+export * from './Panel';
+
+// Layout
+export * from './Layout';
