+++ app/(main)/chat/[questionId]/page.tsx
@@ -1,10 +1,12 @@
 import { Suspense } from 'react';
-import { ChatPageShell } from './_components/ChatPageShell';
-import { ChatAreaServer } from './_components/ChatAreaServer';
-import { ExperienceCardsServer } from './_components/ExperienceCardsServer';
-import { SidePanelClient } from './_components/SidePanelClient';
-import { ChatAreaSkeleton } from './_components/ChatAreaSkeleton';
-import { ExperienceCardsSkeleton } from './_components/ExperienceCardsSkeleton';
+import {
+  ChatPageShell,
+  ChatAreaServer,
+  ChatAreaSkeleton,
+  ExperienceCardsServer,
+  ExperienceCardsSkeleton,
+  SidePanelClient,
+} from './_components';
 
 interface ChatPageProps {
   params: Promise<{
